Improve Durable Objects intro by saying what it does. (#19298)
* Improve Durable Objects intro by saying what it does.

There is a lot of debate about how best to present DOs so that people "get it".

Engineers (like me) usually want to know what a product *does*. This is contrary to the common wisdom that product pages should say what problems a product *solves*. Figuring out how to use a tool to solve a problem is an engineer's job, so as long as they know what it does, they can figure out what it solves. Conversely, if you only tell them what it solves, they then cannot visualize *how* it solves the problem, which is their job, so they get frustrated.

To that end, we should very briefly explain what a Durable Object *does* right on the landing page.