Update dns-records.mdx
The SSL Coverage limitation seems to work fine on first level subdomains with Universal SSL and Full setup.
I understand that limitation happens with Universal SSL and Partial setups, when we have to issue a new certificate for each proxied hostname. I added "Partial setup" with the intention of making it more detailed.